# backend/YanCarz/swagger.json
{
    "openapi":  "3.0.4",
    "info":  {
                 "title":  "YanCarz.API",
                 "version":  "1.0"
             },
    "paths":  {
...
                  "/api/shared/Devise":  {
                                             "get":  {
                                                         "tags":  [
                                                                      "Devise"
                                                                  ],
                                                         "responses":  {
                                                                           "200":  {
                                                                                       "description":  "OK",
                                                                                       "content":  {
                                                                                                       "text/plain":  {
                                                                                                                          "schema":  {
                                                                                                                                         "type":  "array",
                                                                                                                                         "items":  {
                                                                                                                                                       "$ref":  "#/components/schemas/DeviseDto"
                                                                                                                                                   }
                                                                                                                                     }
                                                                                                                      },
                                                                                                       "application/json":  {
                                                                                                                                "schema":  {
                                                                                                                                               "type":  "array",
                                                                                                                                               "items":  {
                                                                                                                                                             "$ref":  "#/components/schemas/DeviseDto"
                                                                                                                                                         }
                                                                                                                                           }
                                                                                                                            },
                                                                                                       "text/json":  {
                                                                                                                         "schema":  {
                                                                                                                                        "type":  "array",
                                                                                                                                        "items":  {
                                                                                                                                                      "$ref":  "#/components/schemas/DeviseDto"
                                                                                                                                                  }
                                                                                                                                    }
                                                                                                                     }
                                                                                                   }
                                                                                   }
                                                                       }
                                                     },
                                             "post":  {
                                                          "tags":  [
                                                                       "Devise"
                                                                   ],
                                                          "requestBody":  {
                                                                              "content":  {
                                                                                              "application/json":  {
                                                                                                                       "schema":  {
                                                                                                                                      "$ref":  "#/components/schemas/DeviseDto"
                                                                                                                                  }
                                                                                                                   },
                                                                                              "text/json":  {
                                                                                                                "schema":  {
                                                                                                                               "$ref":  "#/components/schemas/DeviseDto"
                                                                                                                           }
                                                                                                            },
                                                                                              "application/*+json":  {
                                                                                                                         "schema":  {
                                                                                                                                        "$ref":  "#/components/schemas/DeviseDto"
                                                                                                                                    }
                                                                                                                     }
                                                                                          }
                                                                          },
                                                          "responses":  {
                                                                            "200":  {
                                                                                        "description":  "OK"
                                                                                    }
                                                                        }
                                                      }
                                         },
                  "/api/shared/Devise/{id}":  {
                                                  "get":  {
                                                              "tags":  [
                                                                           "Devise"
                                                                       ],
                                                              "parameters":  [
                                                                                 {
                                                                                     "name":  "id",
                                                                                     "in":  "path",
                                                                                     "required":  true,
                                                                                     "schema":  {
                                                                                                    "type":  "string",
                                                                                                    "format":  "uuid"
                                                                                                }
                                                                                 }
                                                                             ],
                                                              "responses":  {
                                                                                "200":  {
                                                                                            "description":  "OK",
                                                                                            "content":  {
                                                                                                            "text/plain":  {
                                                                                                                               "schema":  {
                                                                                                                                              "$ref":  "#/components/schemas/DeviseDto"
                                                                                                                                          }
                                                                                                                           },
                                                                                                            "application/json":  {
                                                                                                                                     "schema":  {
                                                                                                                                                    "$ref":  "#/components/schemas/DeviseDto"
                                                                                                                                                }
                                                                                                                                 },
                                                                                                            "text/json":  {
                                                                                                                              "schema":  {
                                                                                                                                             "$ref":  "#/components/schemas/DeviseDto"
                                                                                                                                         }
                                                                                                                          }
                                                                                                        }
                                                                                        }
                                                                            }
                                                          },
                                                  "put":  {
                                                              "tags":  [
                                                                           "Devise"
                                                                       ],
                                                              "parameters":  [
                                                                                 {
                                                                                     "name":  "id",
                                                                                     "in":  "path",
                                                                                     "required":  true,
                                                                                     "schema":  {
                                                                                                    "type":  "string",
                                                                                                    "format":  "uuid"
                                                                                                }
                                                                                 }
                                                                             ],
                                                              "requestBody":  {
                                                                                  "content":  {
                                                                                                  "application/json":  {
                                                                                                                           "schema":  {
                                                                                                                                          "$ref":  "#/components/schemas/DeviseDto"
                                                                                                                                      }
                                                                                                                       },
                                                                                                  "text/json":  {
                                                                                                                    "schema":  {
                                                                                                                                   "$ref":  "#/components/schemas/DeviseDto"
                                                                                                                               }
                                                                                                                },
                                                                                                  "application/*+json":  {
                                                                                                                             "schema":  {
                                                                                                                                            "$ref":  "#/components/schemas/DeviseDto"
                                                                                                                                        }
                                                                                                                         }
                                                                                              }
                                                                              },
                                                              "responses":  {
                                                                                "200":  {
                                                                                            "description":  "OK"
                                                                                        }
                                                                            }
                                                          },
                                                  "delete":  {
                                                                 "tags":  [
                                                                              "Devise"
                                                                          ],
                                                                 "parameters":  [
                                                                                    {
                                                                                        "name":  "id",
                                                                                        "in":  "path",
                                                                                        "required":  true,
                                                                                        "schema":  {
                                                                                                       "type":  "string",
                                                                                                       "format":  "uuid"
                                                                                                   }
                                                                                    }
                                                                                ],
                                                                 "responses":  {
                                                                                   "200":  {
                                                                                               "description":  "OK"
                                                                                           }
                                                                               }
                                                             }
                                              },
...
              },
    "components":  {
                       "schemas":  {
...
                                       "DeviseDto":  {
                                                         "type":  "object",
                                                         "properties":  {
                                                                            "id":  {
                                                                                       "type":  "string",
                                                                                       "format":  "uuid"
                                                                                   },
                                                                            "code":  {
                                                                                         "type":  "string",
                                                                                         "nullable":  true
                                                                                     },
                                                                            "name":  {
                                                                                         "type":  "string",
                                                                                         "nullable":  true
                                                                                     }
                                                                        },
                                                         "additionalProperties":  false
                                                     },
...
                                   }
                   },
...
}